chore: move exec and prompt into new run pkg
